backend/requirements.txt
fastapi[standard]
libvirt-python
lxml
psutil